moved code to UserReviewsView
