+++ components/index.ts
@@ -11,6 +11,8 @@ export { default as MenuOption } from "./navigation/MenuOption";
 export { default as ProductHero } from "./views/ProductHero";
 export { default as ServiceProfileView } from "./views/ServiceProfileView";
 export { default as UserProfileView } from "./views/UserProfileView";
+export { default as UserReviewsView } from "./views/UserReviewsView";
+
 // Cards
 export { default as ProductHeroMainCard } from "./cards/ProductHeroMainCard";
 export { default as SearchCard } from "./cards/SearchCard";
